refactor(input-output-row): input/output 공통 위젯 분리
- 공통으로 사용되는 위젯을 InputOutputDetailRow로 분리
- 위젯 내부에서 아이콘 색상, 경로 지정하는 로직 개선(getRowFeature)
- InputOutputRowType enum -> enums 폴더 내부로 이동
#68
