+++ lib/enums/input_output_enums.dart
@@ -1,0 +1,1 @@
+enum InputOutputRowType { input, output, fee }
